Create styles.css
For this 'styles.css' file, I added a few additional styles to better the appearance of the calendar to be easier to see text and bordered sections according to their requirements. 
It has styles for adjusting the font size, as well as; highlighting today's date and a users selected date, adding hover effects to table rows for a straightforward and user friendly design.